feature(as): refactor more tests to use `expectMatch`
